Refactor of SSH/SFTP commands to support optional destination resolution from external source.
