+++ maverick-virtual-session/src/main/java/com/sshtools/vsession/commands/ssh/SshClientOptionsEvaluator.java
@@ -18,7 +18,6 @@
  */
 package com.sshtools.vsession.commands.ssh;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -26,23 +25,11 @@ import java.util.stream.Collectors;
 import org.apache.commons.cli.CommandLine;
 
 import com.sshtools.common.logger.Log;
-import com.sshtools.vsession.commands.ssh.SshClientOptions.CipherSpec;
-import com.sshtools.vsession.commands.ssh.SshClientOptions.Compression;
-import com.sshtools.vsession.commands.ssh.SshClientOptions.IdentityFile;
-import com.sshtools.vsession.commands.ssh.SshClientOptions.LoginName;
-import com.sshtools.vsession.commands.ssh.SshClientOptions.MacSpec;
-import com.sshtools.vsession.commands.ssh.SshClientOptions.Port;
-import com.sshtools.vsession.commands.ssh.SshClientOptions.SecurityLevel;
 
 
-public class SshClientOptionsEvaluator {
+public class SshClientOptionsEvaluator extends AbstractSshOptionsEvaluator {
 
-	static List<SshOptionsResolver> resolvers = new ArrayList<>();
-	
-	public static void addResolver(SshOptionsResolver resolver) {
-		resolvers.add(resolver);
-	}
-	
+
 	public static SshClientArguments evaluate(CommandLine commandLine, String[] originalArguments) {
 		
 		if (Log.isDebugEnabled()) {
@@ -71,29 +58,6 @@ public class SshClientOptionsEvaluator {
 		return arguments;
 	}
 	
-	private static void parseDestination(CommandLine commandLine, SshClientArguments arguments) {
-
-		List<String> commandLineArguments = commandLine.getArgList();
-		String destination = commandLineArguments.get(1);
-		String loginName = null;
-		if (destination.contains("@")) {
-			String[] destinationParts = destination.split("@");
-			loginName = destinationParts[0];
-			destination = destinationParts[1];
-		}
-		
-		arguments.setDestination(destination);
-		arguments.setLoginName(loginName);
-		
-		
-		for(SshOptionsResolver resolver : resolvers) {
-			if(resolver.resolveDestination(destination, arguments)) {
-				break;
-			}
-		}
-	}
-	
-	
 	private static void parseCommand(String[] originalArguments, SshClientArguments arguments) {
 		int indexTillSshClientCommandFound = SshClientOptionsExtractor.extractSshCommandLineFromExecuteCommand(originalArguments);
 		if (originalArguments.length == (indexTillSshClientCommandFound + 1)) {
@@ -104,73 +68,6 @@ public class SshClientOptionsEvaluator {
 		String[] executeCommand = Arrays.copyOfRange(originalArguments, indexTillSshClientCommandFound + 1, originalArguments.length);
 		String command = String.join(" ", executeCommand);
 		arguments.setCommand(command);
-	}
-	
-	private static void parsePort(CommandLine commandLine, SshClientArguments arguments) {
-		int port = 22;
-		
-		if (commandLine.hasOption(Port.PORT_OPTION)) {
-			String portValue = commandLine.getOptionValue(Port.PORT_OPTION);
-			try {
-				port = Integer.parseInt(portValue);
-			} catch (Exception e) {
-				port = 22;
-			}
-		}
-		
-		arguments.setPort(port);
-	}
-	
-	private static void parseLoginName(CommandLine commandLine, SshClientArguments arguments) {
-
-		if (commandLine.hasOption(LoginName.LOGIN_NAME_OPTION)) {
-			arguments.setLoginName(commandLine.getOptionValue(LoginName.LOGIN_NAME_OPTION));
-		}
-		
-	}
-	
-	private static void parseIdentityFilename(CommandLine commandLine, SshClientArguments arguments) {
-
-		if (commandLine.hasOption(IdentityFile.IDENTITY_FILE_OPTION)) {
-			arguments.setIdentityFile(commandLine.getOptionValue(IdentityFile.IDENTITY_FILE_OPTION));
-		}
-		
-	}
-	
-	private static void parseCiphers(CommandLine commandLine, SshClientArguments arguments) {
-
-		if (commandLine.hasOption(CipherSpec.CIPHER_SPEC_OPTION)) {
-			String[] cipherSpecParts = commandLine.getOptionValues(CipherSpec.CIPHER_SPEC_OPTION);
-			String[] finalValues = CommandUtil.toStringFromCsvs(cipherSpecParts);
-			arguments.setCiphers(finalValues);
-		}
-		
-	}
-	
-	private static void parseMacs(CommandLine commandLine, SshClientArguments arguments) {
-
-		if (commandLine.hasOption(MacSpec.MAC_SPEC_OPTION)) {
-			String[] macSpecParts = commandLine.getOptionValues(MacSpec.MAC_SPEC_OPTION);
-			String[] finalValues = CommandUtil.toStringFromCsvs(macSpecParts);
-			arguments.setHmacs(finalValues);
-		}
-		
-	}
-	
-	private static void parseSecurityLevel(CommandLine commandLine, SshClientArguments arguments) {
-
-		if (commandLine.hasOption(SecurityLevel.SECURITY_LEVEL_OPTION)) {
-			String securityLevel = commandLine.getOptionValue(SecurityLevel.SECURITY_LEVEL_OPTION);
-			arguments.setSecurityLevel(securityLevel);
-		}
-		
-	}
-	
-	private static void parseCompression(CommandLine commandLine, SshClientArguments arguments) {
-		
-		if (commandLine.hasOption(Compression.COMPRESSION_OPTION)) {
-			arguments.setCompression(true);
-		}
 	}
 	
 }
